refactor: modify duplicated file, component names
